+++ lib/src/theme/filter_list_theme.dart
@@ -9,27 +9,28 @@ class FilterListTheme extends InheritedWidget {
   /// Builds a [FilterListTheme].
   const FilterListTheme({
     Key? key,
-    required this.data,
+    required this.theme,
     required Widget child,
   }) : super(key: key, child: child);
 
   /// The [FilterListThemeData] styling for this theme.
-  final FilterListThemeData data;
+  final FilterListThemeData theme;
 
   @override
-  bool updateShouldNotify(FilterListTheme oldWidget) => data != oldWidget.data;
+  bool updateShouldNotify(FilterListTheme oldWidget) =>
+      theme != oldWidget.theme;
 
   /// Retrieves the [FilterListThemeData] from the closest ancestor
   /// [FilterListTheme] widget.
   static FilterListThemeData of(BuildContext context) {
-    final theme = context.dependOnInheritedWidgetOfExactType<FilterListTheme>();
-
+    FilterListTheme? theme =
+        context.dependOnInheritedWidgetOfExactType<FilterListTheme>();
     assert(
       theme != null,
       'You must have a FilterListTheme widget at the top of your widget tree',
     );
 
-    return theme!.data;
+    return theme!.theme;
   }
 }
 
@@ -40,20 +41,19 @@ class FilterListThemeData with Diagnosticable {
   factory FilterListThemeData({
     required BuildContext context,
     Brightness? brightness,
-    IconThemeData? primaryIconTheme,
     ChoiceChipThemeData? choiceChipTheme,
     HeaderThemeData? headerTheme,
     ControlButtonBarThemeData? controlButtonBarTheme,
+    double? borderRadius,
+    WrapAlignment? wrapAlignment,
+    WrapCrossAlignment? wrapCrossAxisAlignment,
+    double? wrapSpacing,
+    Color? backgroundColor,
   }) {
     // Use the given brightness, or a default
     final _brightness = brightness ?? Brightness.light;
     // Determine dark or light
     final isDark = _brightness == Brightness.dark;
-
-    // Use the given primaryIconTheme or a default.
-    primaryIconTheme ??= IconThemeData(
-      color: isDark ? const Color(0xff959595) : const Color(0xff757575),
-    );
 
     choiceChipTheme = choiceChipTheme ?? ChoiceChipThemeData.light(context);
 
@@ -62,12 +62,31 @@ class FilterListThemeData with Diagnosticable {
     controlButtonBarTheme =
         controlButtonBarTheme ?? ControlButtonBarThemeData.light(context);
 
+    /// Border radius of filter dialog
+    borderRadius ??= 20;
+
+    /// Background color of filter dialog
+    backgroundColor ??= isDark ? Colors.black : Colors.white;
+
+    /// Wrap alignment of filter dialog
+    wrapAlignment ??= WrapAlignment.start;
+
+    /// Wrap cross alignment of filter dialog
+    wrapCrossAxisAlignment ??= WrapCrossAlignment.start;
+
+    /// Wrap spacing of filter dialog
+    wrapSpacing ??= 0.0;
+
     return FilterListThemeData.raw(
       brightness: _brightness,
-      primaryIconTheme: primaryIconTheme,
       choiceChipTheme: choiceChipTheme,
       headerTheme: headerTheme,
       controlBarButtonTheme: controlButtonBarTheme,
+      borderRadius: borderRadius,
+      wrapAlignment: wrapAlignment,
+      wrapCrossAxisAlignment: wrapCrossAxisAlignment,
+      wrapSpacing: wrapSpacing,
+      backgroundColor: backgroundColor,
     );
   }
 
@@ -80,18 +99,45 @@ class FilterListThemeData with Diagnosticable {
       FilterListThemeData(brightness: Brightness.dark, context: context);
 
   /// Raw [FilterListThemeData] initialization.
-  const FilterListThemeData.raw(
-      {required this.brightness,
-      required this.primaryIconTheme,
-      required this.choiceChipTheme,
-      required this.headerTheme,
-      required this.controlBarButtonTheme});
+  const FilterListThemeData.raw({
+    required this.brightness,
+    required this.choiceChipTheme,
+    required this.headerTheme,
+    required this.controlBarButtonTheme,
+    required this.borderRadius,
+    required this.wrapAlignment,
+    required this.wrapCrossAxisAlignment,
+    required this.wrapSpacing,
+    required this.backgroundColor,
+  });
 
   /// The [Brightness] of this theme.
   final Brightness brightness;
 
-  /// The primary icon theme
-  final IconThemeData primaryIconTheme;
+  /// Border radius of the filter dialog
+  final double borderRadius;
+
+  /// How the choice chip within a run should be placed in the main axis.
+  /// For example, if [wrapSpacing] is [WrapAlignment.center], the choice chip in each run are grouped together in the center of their run in the main axis.
+  ///
+  /// Defaults to [WrapAlignment.start].
+  final WrapAlignment wrapAlignment;
+
+  /// How the choice chip within a run should be aligned relative to each other in the cross axis.
+  ///For example, if this is set to [WrapCrossAlignment.end], and the [direction] is [Axis.horizontal], then the choice chip within each run will have their bottom edges aligned to the bottom edge of the run.
+  ///
+  ///Defaults to [WrapCrossAlignment.start].
+  final WrapCrossAlignment wrapCrossAxisAlignment;
+
+  ///How much space to place between choice chip in a run in the main axis.
+  ///For example, if [wrapSpacing] is 10.0, the choice chip will be spaced at least 10.0 logical pixels apart in the main axis.
+  ///If there is additional free space in a run (e.g., because the wrap has a minimum size that is not filled or because some runs are longer than others), the additional free space will be allocated according to the [alignment].
+  ///
+  ///Defaults to 0.0.
+  final double wrapSpacing;
+
+  /// Background color of the filter list
+  final Color? backgroundColor;
 
   /// {@macro choice_chip_theme}
   final ChoiceChipThemeData choiceChipTheme;
@@ -107,8 +153,17 @@ class FilterListThemeData with Diagnosticable {
     super.debugFillProperties(properties);
     properties
       ..add(EnumProperty<Brightness>('brightness', brightness))
-      ..add(DiagnosticsProperty<IconThemeData>(
-          'primaryIconTheme', primaryIconTheme))
+      ..add(DiagnosticsProperty<double>('borderRadius', borderRadius))
+      ..add(DiagnosticsProperty<Color>('backgroundColor', backgroundColor))
+      ..add(DiagnosticsProperty<ChoiceChipThemeData>(
+          'choiceChipTheme', choiceChipTheme))
+      ..add(DiagnosticsProperty<HeaderThemeData>('headerTheme', headerTheme))
+      ..add(DiagnosticsProperty<ControlButtonBarThemeData>(
+          'controlButtonBarTheme', controlBarButtonTheme))
+      ..add(DiagnosticsProperty<WrapAlignment>('wrapAlignment', wrapAlignment))
+      ..add(DiagnosticsProperty<WrapCrossAlignment>(
+          'wrapCrossAxisAlignment', wrapCrossAxisAlignment))
+      ..add(DiagnosticsProperty<double>('wrapSpacing', wrapSpacing))
       ..add(DiagnosticsProperty<ChoiceChipThemeData>(
           'choiceChipTheme', choiceChipTheme))
       ..add(DiagnosticsProperty<HeaderThemeData>('headerTheme', headerTheme))